// js/stat-evaluation-player/src/generated/HalfFlipStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledCounts } from "./LabeledCounts.ts";

export type HalfFlipStats = { count: number, high_confidence_count: number, is_last_half_flip: boolean, last_half_flip_time: number | null, last_half_flip_frame: number | null, time_since_last_half_flip: number | null, frames_since_last_half_flip: number | null, last_quality: number | null, best_quality: number, cumulative_quality: number, labeled_event_counts?: LabeledCounts, };