// js/stat-evaluation-player/src/generated/WavedashEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type WavedashEvent = { time: number, frame: number, player: RemoteIdTs, is_team_0: boolean, dodge_time: number, dodge_frame: number, time_since_dodge: number, dodge_position: [number, number, number], landing_position: [number, number, number], start_speed: number, landing_speed: number, horizontal_speed_gain: number, landing_uprightness: number, confidence: number, };