// js/stat-evaluation-player/src/generated/WavedashStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledCounts } from "./LabeledCounts.ts";

export type WavedashStats = { count: number, high_confidence_count: number, is_last_wavedash: boolean, last_wavedash_time: number | null, last_wavedash_frame: number | null, time_since_last_wavedash: number | null, frames_since_last_wavedash: number | null, last_quality: number | null, best_quality: number, cumulative_quality: number, labeled_event_counts?: LabeledCounts, };